fixed uint8_t cin bug

+++ src/workbench.h
@@ -14,18 +14,18 @@ class WorkBench
 {
 private:
 public:
-    WorkBench(uint8_t type, double x, double y);
-    WorkBench(uint8_t type, double x, double y, int product_frames_remained,
-              uint8_t material_status, bool product_status);
+    WorkBench(uint16_t type, double x, double y);
+    WorkBench(uint16_t type, double x, double y, int product_frames_remained,
+              uint16_t material_status, bool product_status);
     double ETA();                              // 剩余生产时间（秒）
     [[nodiscard]] bool isFree(int index) const;// 查看指定原材料格状态
     [[nodiscard]] bool isReady() const;        // 查看产品格状态
     bool needRawMaterial(int index);           // 原材料格里是否有index这种类型的原料
 
-    uint8_t type;               // 工作台类型
+    uint16_t type;              // 工作台类型
     Point coordinate;           //坐标
     int product_frames_remained;// 剩余生产时间
-    uint8_t material_status;    // 原材料格状态
+    uint16_t material_status;   // 原材料格状态
     bool product_status;        // 产品格状态
     int id;
 };
